fixed bugs.

+++ src/se/kjellstrand/onemillion/ExchangeRates.java
@@ -1,10 +1,12 @@
+
 package se.kjellstrand.onemillion;
 
 import java.util.Map;
 
 public class ExchangeRates {
-	
-	public String base;
-	public Map<String,Float> rates;
+
+    public String base;
+
+    public Map<String, Double> rates;
 
 }
